+++ tests/_core/solver/_strategies/_initialization/test_base.py
@@ -13,6 +13,8 @@ from max_div._core.solver._strategies import InitializationStrategy
     "factory_method",
     [
         InitializationStrategy.fast,
+        partial(InitializationStrategy.farthest_point, top_k=4),
+        partial(InitializationStrategy.most_feasible, max_iter=100),
         partial(InitializationStrategy.random_one_shot, ignore_constraints=False),
         partial(InitializationStrategy.random_one_shot, ignore_constraints=True),
         partial(InitializationStrategy.random_batched, b=2, ignore_constraints=False),
@@ -22,6 +24,8 @@ from max_div._core.solver._strategies import InitializationStrategy
     ],
     ids=[
         "fast",
+        "farthest_point",
+        "most_feasible",
         "random_one_shot_1",
         "random_one_shot_2",
         "random_batched_1",